18.12.2021 все сущности созданы и работают (но коряво).

+++ views/books/index.php
@@ -3,11 +3,14 @@
 
     use App\config\DB_connect;
 
+    /** @var array $model */
+    /** @var array $access */
+    /** @var array $connectBooks */
+    /** @var array $readersTicket */
+
     \App\core\Breadcrumb::add_current('/user/index', 'Книги');
     echo \App\core\Breadcrumb::out();
-    $entityManagerClass = new DB_connect();
-    $entityManager = $entityManagerClass->connect();
-    /** @var array $model */
+
     ?>
 </div>
 
@@ -25,22 +28,38 @@
         </thead>
         <?php foreach ($model as $models) : ?>
             <tr>
-
-                <td align="left"><?php echo $models->getNameBooks(); ?></td>
+                <td align="left"><a
+                            href="/books/view?id=<?php echo $models->getIdBooks() ?>"><?php echo $models->getNameBooks(); ?></a>
+                </td>
                 <td align="left"><?php echo $models->getAuthor(); ?></td>
                 <td align="left"><?php echo $models->getDatePublication(); ?></td>
                 <td>
-                    <a href="/books/view?id=<?php echo $models->getIdBooks(); ?>"" class="btn btn-outline-info
-                    btn-sm"><i class="bi bi-eye"></i></a>
-                    <a href="/books/update?id=<?php echo $models->getIdBooks(); ?>"
-                       class="btn btn-outline-primary btn-sm"><i class="bi bi-pencil"></i></a>
-                    <a href="/books/delete?id=<?php echo $models->getIdBooks(); ?>"
-                       class="btn btn-outline-danger btn-sm"><i class="bi bi-trash"></i>
+                    <?php if ($access->getRole('Администратор') || $access->getRole('Сотрудник библиотеки')) { ?>
+                        <a href="/books/view?id=<?php echo $models->getIdBooks(); ?>"" class="btn btn-outline-info btn-sm">
+                        <i class="bi bi-eye"></i></a>
+                        <a href="/books/update?id=<?php echo $models->getIdBooks(); ?>"
+                           class="btn btn-outline-primary btn-sm"><i class="bi bi-pencil"></i></a>
+                        <a href="/books/delete?id=<?php echo $models->getIdBooks(); ?>"
+                           class="btn btn-outline-danger btn-sm"><i class="bi bi-trash"></i></a>
+                        <?php
+                    } else {
+                        $readersTicketBlock = $readersTicket->getOneUser();
+                        $books = $connectBooks->getAvailability($models->getIdBooks());
+                        if (empty($readersTicketBlock)) {
+                            if (!empty($books)) {
+                                echo ' <a class="btn btn-success btn-sm">Добавлена в Ваши книги</a>';
+                            } else echo '<a href="/books-user/add?id=' . $models->getIdBooks() . ' "class="btn btn-outline-info btn-sm">Получить книгу</a>';
+                        } else {
+                            echo '<a href="/books-user/index" class="btn btn-danger btn-sm">Вы заблокированы</a>';
+                        }
+                    }
+                    ?>
                 </td>
             </tr>
         <?php endforeach; ?>
     </table>
-    <div class="col-md-2 text-right">
-        <a href="/books/create" class="btn btn-success">Добавить</a>
-    </div>
-
+    <?php if ($access->getRole('Администратор') || $access->getRole('Сотрудник библиотеки')) : ?>
+        <div class="col-md-2 text-right">
+            <a href="/books/create" class="btn btn-success">Добавить</a>
+        </div>
+    <?php endif; ?>
